iter1: add handling of withdrawals

+++ migrations/000003_create_balance_table.up.sql
@@ -1,8 +1,8 @@
 CREATE TABLE IF NOT EXISTS balance (
     id SERIAL PRIMARY KEY,
     owner_id INTEGER NOT NULL,
-    current DECIMAL(15, 2) NOT NULL DEFAULT 0,
-    withdrawn DECIMAL(15, 2) NOT NULL DEFAULT 0,
+    current INTEGER NOT NULL DEFAULT 0,
+    withdrawn INTEGER NOT NULL DEFAULT 0,
     
     CONSTRAINT fk_user
         FOREIGN KEY(owner_id)
